backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.3
pydantic-settings==2.6.1
numpy==1.26.4
pandas>=2.0.0
python-dotenv>=1.0.0
yfinance>=0.2.0
openai>=1.0.0
pytest>=8.0.0
pytest-cov>=4.0.0